backend/pyproject.toml
[project]
name = "anordating"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiogram==3.17.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "alembic==1.14.0",
    "alembic-postgresql-enum==1.5.0",
    "annotated-types==0.7.0",
    "async-timeout==5.0.1",
    "asyncpg==0.30.0",
    "attrs==24.3.0",
    "babel==2.16.0",
    "certifi==2024.12.14",
    "fastapi[standard,uvicorn]>=0.115.12",
    "frozenlist==1.5.0",
    "googlemaps>=4.10.0",
    "greenlet>=3.1.1",
    "idna==3.10",
    "magic-filter==1.0.12",
    "mako==1.3.8",
    "markupsafe==3.0.2",
    "motor>=3.7.0",
    "multidict==6.1.0",
    "propcache==0.2.1",
    "psycopg-binary==3.2.3",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pymongo[srv]>=4.11",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.36",
    "typing-extensions==4.12.2",
    "uvicorn>=0.34.1",
    "yarl==1.18.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "watchdog[watchmedo]>=6.0.0",
]
